requirements.txt
# Python 3.11+ is required
# Core dependencies
pybit>=5.0.0
pandas>=2.0.0  # For Python 3.11 compatibility
numpy>=1.24.0  # For Python 3.11 compatibility
python-telegram-bot>=20.0
pandas-ta>=0.3.14b0
python-dotenv>=1.0.0
psutil # For system monitoring

# Web interface dependencies
flask>=2.3.0  # For Python 3.11 compatibility
flask-socketio>=5.3.0  # For Python 3.11 compatibility
flask-login>=0.6.2
flask-wtf>=1.1.0
eventlet>=0.33.0  # Required for Socket.IO with Python 3.11
waitress>=2.1.2  # Production WSGI server
simple-websocket>=0.10.0
